requirements.txt
websockets==11.0.3
aiohttp==3.9.5
pyyaml==6.0.1
prometheus-client==0.20.0
python-dotenv==1.0.1
redis==5.0.7
PyJWT==2.9.0
